Added reindex apis
