+++ llm-server/shared/utils/opencopilot_utils/__init__.py
@@ -6,3 +6,4 @@ from .store_type import StoreType
 from .interfaces import StoreOptions
 from .get_llm import get_llm
 from .get_shared_filepath import get_file_path
+from .get_vector_store import get_vector_store
