+++ llm-server/workers/tasks/__init__.py
@@ -1,2 +1,3 @@
 from workers.tasks.process_pdfs import process_pdf
 from workers.tasks.web_crawl import web_crawl
+from workers.tasks.reindex_swagger import reindex_apis
